requirements.txt
altair==4.2.2
appdirs==1.4.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.39.3
frozendict==2.3.7
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
lime==0.2.0.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multitasking==0.0.11
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pyzmq==25.0.2
requests==2.28.2
rich==13.3.4
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.21.0
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
yfinance==0.2.18
zipp==3.15.0